Refine utest by AI.
